refactor: move bucketStorageHelper to utils
also updated comment regarding broken tests in readFromStoreInBuckets
