--- src/context/authStore/index.ts
@@ -1,1 +1,0 @@
-export * from "./authStore";
